pyproject.toml
[project]
name = "python-graphql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "sqlalchemy>=2.0.44",
    "strawberry-graphql>=0.287.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "pylint>=4.0.3",
    "pytest>=9.0.1",
    "tox>=4.32.0",
]
